// types/types.ts
// Data types likely to be generated by various systems, added manually here.

export interface Category {
  products: Product[];
  title: string;
}

export interface Product {
  color: {
    hex: string;
    title: string;
    value: string;
  };
  id: string;
  name: string;
  price: {
    title: string;
    value: number;
    currency: string;
  };
  savingsText?: string | null;
  productTag?: string;
  images: {
    product: Image[];
  };
}

export interface Image {
  alt?: string;
  url: string;
}